Fixed bug in insertion rank + minor formatting.
Conditional to check if a value was greater than the other
was flawed. Integer overflow could occur if two values were
close to either end of the integer length which made if
statement redundant. Probably not picked up until now as the
calculation of the infinite temperature trace was recently
optimised to use the fact that the list should have been
sorted.